Havent-me fixat a la presentació que el servidor esporàdic donava uns resultats molt extranys i extramedament lent. L'he revisat, s'han detectat errors i els he corregits. També s'ha guanyat molt amb eficiència. El tempsde resposta ara amb l'esporàdic és molt millor, gairabé igualanat al Background Server.

+++ trunk/scheduling-algorithms/src/scheduling/algorithms/DeferrableServer.java
@@ -18,7 +18,7 @@ public class DeferrableServer extends Server {
     }
 
     @Override
-    public int scheduleAperiodicTaskGroup(Summary summary) {
+    public double scheduleAperiodicTaskGroup(Summary summary) {
         //throw new UnsupportedOperationException("Not supported yet.");
         ListIterator<Node> inode = summary.getSummaryListIterator();
         Server.getAperiodicTaskGroup().sortByArrivalTime();
@@ -30,7 +30,7 @@ public class DeferrableServer extends Server {
         int ciclo = 0;                                  //Ciclo Actual
         int tiempoActual = 0;                           //tiempo Actual
         int tiempoPendiente = 0;                        //Tiempo Pendiente
-        int tiempoTotalTareas = 0;                      //Tiempo total para calcular el tiempo de respuesta
+        double tiempoTotalTareas = 0;                      //Tiempo total para calcular el tiempo de respuesta
         int ntareas = 0;       
         boolean handled = false;
